+++ views/issues_views.xml
@@ -30,7 +30,7 @@
                                 <field name="type"/>
                                 <field name="priority"/>
                                 <field name="assignee_id"/>
-
+                                <field name="resolution"/>
                             </group>
                         </group>
                         <notebook>
@@ -97,6 +97,7 @@
                     <field name="type"/>
                     <field name="priority"/>
                     <field name="status"/>
+                    <field name="resolution"/>
                     <field name="function_id"/>
                     <field name="assignee_id"/>
                     <field name="create_date"/>
